fix: address review feedback on PR #85
