+++ src/lib/mode/aram.ts
@@ -7,14 +7,9 @@
 
 import type { GameState } from "../game-state/types";
 import type { LoadedGameData } from "../data-ingest";
-import type {
-  GameMode,
-  ModeContext,
-  PlayerModeContext,
-  TeamComposition,
-} from "./types";
+import type { GameMode, ModeContext, PlayerModeContext } from "./types";
 import { GAME_MODE_ARAM } from "./types";
-import { filterItemsByMode } from "./utils";
+import { filterItemsByMode, buildTeamComposition } from "./utils";
 
 export const aramMode: GameMode = {
   id: "aram",
@@ -68,13 +63,3 @@ export const aramMode: GameMode = {
     };
   },
 };
-
-function buildTeamComposition(players: PlayerModeContext[]): TeamComposition {
-  const classCounts: Record<string, number> = {};
-  for (const player of players) {
-    for (const tag of player.tags) {
-      classCounts[tag] = (classCounts[tag] ?? 0) + 1;
-    }
-  }
-  return { players, classCounts };
-}
